(Sprinkler) 細かいクラスを別ファイルに分離 #58
