+++ PRG/Tools/Sprinkler/Sprinkler/InputConverter.cs
@@ -1,9 +1,8 @@
-﻿using System;
+﻿using Midi;
+using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 using System.Windows.Forms;
-using Midi;
 
 namespace Sprinkler
 {
